[F] fitText again after fonts load
When a font isn't cached locally, it takes time to load, and the
original code may cause timing issues with it. Fitting the text
first and then loading the font may result in a smaller font size
than expected. You can disable cache and enable throttling in the
Network tab in F12, then refresh the page to see this bug.

https://t.me/c/1656800117/30919